+++ src/meta/hca.c
@@ -3,6 +3,7 @@
 #include "../coding/coding.h"
 #include "../coding/hca_decoder_clhca.h"
 #include "../util/channel_mappings.h"
+#include "../util/companion_files.h"
 
 #ifdef VGM_DEBUG_OUTPUT
   //#define HCA_BRUTEFORCE
