+++ src/streamfile.c
@@ -2,20 +2,12 @@
 #include "util.h"
 #include "vgmstream.h"
 #include "util/reader_sf.h"
+#include "util/paths.h"
 #include <string.h>
 
 /* for dup/fdopen in some systems */
 #ifndef _MSC_VER
     #include <unistd.h>
-#endif
-
-//TODO: move
-#ifndef DIR_SEPARATOR
-    #if defined (_WIN32) || defined (WIN32)
-        #define DIR_SEPARATOR '\\'
-    #else
-        #define DIR_SEPARATOR '/'
-    #endif
 #endif
 
 /* For (rarely needed) +2GB file support we use fseek64/ftell64. Those are usually available
@@ -1205,162 +1197,6 @@ size_t read_string_utf16be(char* buf, size_t buf_size, off_t offset, STREAMFILE*
 
 /* ************************************************************************* */
 
-size_t read_key_file(uint8_t* buf, size_t buf_size, STREAMFILE* sf) {
-    char keyname[PATH_LIMIT];
-    char filename[PATH_LIMIT];
-    const char *path, *ext;
-    STREAMFILE* sf_key = NULL;
-    size_t keysize;
-
-    get_streamfile_name(sf, filename, sizeof(filename));
-
-    if (strlen(filename)+4 > sizeof(keyname)) goto fail;
-
-    /* try to open a keyfile using variations */
-    {
-        ext = strrchr(filename,'.');
-        if (ext!=NULL) ext = ext+1;
-
-        path = strrchr(filename, DIR_SEPARATOR);
-        if (path!=NULL) path = path+1;
-
-        /* "(name.ext)key" */
-        strcpy(keyname, filename);
-        strcat(keyname, "key");
-        sf_key = sf->open(sf, keyname, STREAMFILE_DEFAULT_BUFFER_SIZE);
-        if (sf_key) goto found;
-
-        /* "(name.ext)KEY" */
-        /*
-        strcpy(keyname+strlen(keyname)-3,"KEY");
-        sf_key = sf->open(sf, keyname, STREAMFILE_DEFAULT_BUFFER_SIZE);
-        if (sf_key) goto found;
-        */
-
-
-        /* "(.ext)key" */
-        if (path) {
-            strcpy(keyname, filename);
-            keyname[path-filename] = '\0';
-            strcat(keyname, ".");
-        } else {
-            strcpy(keyname, ".");
-        }
-        if (ext) strcat(keyname, ext);
-        strcat(keyname, "key");
-        sf_key = sf->open(sf, keyname, STREAMFILE_DEFAULT_BUFFER_SIZE);
-        if (sf_key) goto found;
-
-        /* "(.ext)KEY" */
-        /*
-        strcpy(keyname+strlen(keyname)-3,"KEY");
-        sf_key = sf->open(sf, keyname, STREAMFILE_DEFAULT_BUFFER_SIZE);
-        if (sf_key) goto found;
-        */
-
-        goto fail;
-    }
-
-found:
-    keysize = get_streamfile_size(sf_key);
-    if (keysize > buf_size) goto fail;
-
-    if (read_streamfile(buf, 0, keysize, sf_key) != keysize)
-        goto fail;
-
-    close_streamfile(sf_key);
-    return keysize;
-
-fail:
-    close_streamfile(sf_key);
-    return 0;
-}
-
-STREAMFILE* read_filemap_file(STREAMFILE* sf, int file_num) {
-    return read_filemap_file_pos(sf, file_num, NULL);
-}
-
-STREAMFILE* read_filemap_file_pos(STREAMFILE* sf, int file_num, int* p_pos) {
-    char filename[PATH_LIMIT];
-    off_t txt_offset, file_size;
-    STREAMFILE* sf_map = NULL;
-    int file_pos = 0;
-
-    sf_map = open_streamfile_by_filename(sf, ".txtm");
-    if (!sf_map) goto fail;
-
-    get_streamfile_filename(sf, filename, sizeof(filename));
-
-    txt_offset = read_bom(sf_map);
-    file_size = get_streamfile_size(sf_map);
-
-    /* read lines and find target filename, format is (filename): value1, ... valueN */
-    while (txt_offset < file_size) {
-        char line[0x2000];
-        char key[PATH_LIMIT] = { 0 }, val[0x2000] = { 0 };
-        int ok, bytes_read, line_ok;
-
-        bytes_read = read_line(line, sizeof(line), txt_offset, sf_map, &line_ok);
-        if (!line_ok) goto fail;
-
-        txt_offset += bytes_read;
-
-        /* get key/val (ignores lead/trailing spaces, stops at comment/separator) */
-        ok = sscanf(line, " %[^\t#:] : %[^\t#\r\n] ", key, val);
-        if (ok != 2) { /* ignore line if no key=val (comment or garbage) */
-            /* better way? */
-            if (strcmp(line, "#@reset-pos") == 0) {
-                file_pos = 0;
-            }
-            continue;
-        }
-
-        if (strcmp(key, filename) == 0) {
-            int n;
-            char subval[PATH_LIMIT];
-            const char* current = val;
-            int i;
-
-            for (i = 0; i <= file_num; i++) {
-                if (current[0] == '\0')
-                    goto fail;
-
-                ok = sscanf(current, " %[^\t#\r\n,]%n ", subval, &n);
-                if (ok != 1)
-                    goto fail;
-
-                if (i == file_num) {
-                    if (p_pos) *p_pos = file_pos;
-
-                    close_streamfile(sf_map);
-                    return open_streamfile_by_filename(sf, subval);
-                }
-
-                current += n;
-                if (current[0] == ',')
-                    current++;
-            }
-        }
-        file_pos++;
-    }
-
-fail:
-    close_streamfile(sf_map);
-    return NULL;
-}
-
-void fix_dir_separators(char* filename) {
-    char c;
-    int i = 0;
-    while ((c = filename[i]) != '\0') {
-        if ((c == '\\' && DIR_SEPARATOR == '/') || (c == '/' && DIR_SEPARATOR == '\\'))
-            filename[i] = DIR_SEPARATOR;
-        i++;
-    }
-}
-
-/* ************************************************************************* */
-
 int check_extensions(STREAMFILE* sf, const char* cmp_exts) {
     char filename[PATH_LIMIT];
     const char* ext = NULL;
@@ -1389,73 +1225,6 @@ int check_extensions(STREAMFILE* sf, const char* cmp_exts) {
     } while (cmp_ext != NULL);
 
     return 0;
-}
-
-/* ************************************************************************* */
-
-/**
- * Find a chunk starting from an offset, and save its offset/size (if not NULL), with offset after id/size.
- * Works for chunked headers in the form of "chunk_id chunk_size (data)"xN  (ex. RIFF).
- * The start_offset should be the first actual chunk (not "RIFF" or "WAVE" but "fmt ").
- * "full_chunk_size" signals chunk_size includes 4+4+data.
- *
- * returns 0 on failure
- */
-static int find_chunk_internal(STREAMFILE* sf, uint32_t chunk_id, off_t start_offset, size_t max_size, int full_chunk_size, off_t *out_chunk_offset, size_t *out_chunk_size, int big_endian_type, int big_endian_size, int zero_size_end) {
-    int32_t (*read_32bit_type)(off_t,STREAMFILE*) = big_endian_type ? read_32bitBE : read_32bitLE;
-    int32_t (*read_32bit_size)(off_t,STREAMFILE*) = big_endian_size ? read_32bitBE : read_32bitLE;
-    off_t offset, max_offset;
-    size_t file_size = get_streamfile_size(sf);
-
-    if (max_size == 0)
-        max_size = file_size;
-
-    offset = start_offset;
-    max_offset = offset + max_size;
-    if (max_offset > file_size)
-        max_offset = file_size;
-
-
-    /* read chunks */
-    while (offset < max_offset) {
-        uint32_t chunk_type = read_32bit_type(offset + 0x00,sf);
-        uint32_t chunk_size = read_32bit_size(offset + 0x04,sf);
-
-        if (chunk_type == 0xFFFFFFFF || chunk_size == 0xFFFFFFFF)
-            return 0;
-
-        if (chunk_type == chunk_id) {
-            if (out_chunk_offset) *out_chunk_offset = offset + 0x08;
-            if (out_chunk_size) *out_chunk_size = chunk_size;
-            return 1;
-        }
-
-        /* empty chunk with 0 size, seen in some formats (XVAG uses it as end marker, Wwise doesn't) */
-        if (chunk_size == 0 && zero_size_end)
-            return 0;
-
-        offset += full_chunk_size ? chunk_size : 0x08 + chunk_size;
-    }
-
-    return 0;
-}
-int find_chunk_be(STREAMFILE* sf, uint32_t chunk_id, off_t start_offset, int full_chunk_size, off_t *out_chunk_offset, size_t *out_chunk_size) {
-    return find_chunk(sf, chunk_id, start_offset, full_chunk_size, out_chunk_offset, out_chunk_size, 1, 0);
-}
-int find_chunk_le(STREAMFILE* sf, uint32_t chunk_id, off_t start_offset, int full_chunk_size, off_t *out_chunk_offset, size_t *out_chunk_size) {
-    return find_chunk(sf, chunk_id, start_offset, full_chunk_size, out_chunk_offset, out_chunk_size, 0, 0);
-}
-int find_chunk(STREAMFILE* sf, uint32_t chunk_id, off_t start_offset, int full_chunk_size, off_t *out_chunk_offset, size_t *out_chunk_size, int big_endian_size, int zero_size_end) {
-    return find_chunk_internal(sf, chunk_id, start_offset, 0, full_chunk_size, out_chunk_offset, out_chunk_size, 1, big_endian_size, zero_size_end);
-}
-int find_chunk_riff_le(STREAMFILE* sf, uint32_t chunk_id, off_t start_offset, size_t max_size, off_t *out_chunk_offset, size_t *out_chunk_size) {
-    return find_chunk_internal(sf, chunk_id, start_offset, max_size, 0, out_chunk_offset, out_chunk_size, 1, 0, 0);
-}
-int find_chunk_riff_be(STREAMFILE* sf, uint32_t chunk_id, off_t start_offset, size_t max_size, off_t *out_chunk_offset, size_t *out_chunk_size) {
-    return find_chunk_internal(sf, chunk_id, start_offset, max_size, 0, out_chunk_offset, out_chunk_size, 1, 1, 0);
-}
-int find_chunk_riff_ve(STREAMFILE* sf, uint32_t chunk_id, off_t start_offset, size_t max_size, off_t *out_chunk_offset, size_t *out_chunk_size, int big_endian) {
-    return find_chunk_internal(sf, chunk_id, start_offset, max_size, 0, out_chunk_offset, out_chunk_size, big_endian, big_endian, 0);
 }
 
 /* ************************************************************************* */
